+++ src/app/shared/components/create-payment-link-form/create-payment-link-form.component.ts
@@ -3,7 +3,6 @@ import { Validators } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { FormBuilder } from '@ngneat/reactive-forms';
 import { FbGroupConfig } from '@ngneat/reactive-forms/lib/formBuilder';
-import { ControlsValue } from '@ngneat/reactive-forms/lib/types';
 import { TranslocoService } from '@ngneat/transloco';
 import { UntilDestroy } from '@ngneat/until-destroy';
 import { ComponentChanges } from '@rbkmoney/utils';
@@ -11,7 +10,7 @@ import { provideValueAccessor } from '@s-libs/ng-core';
 
 import { BankCard, PaymentMethod, PaymentTerminal } from '@dsh/api-codegen/capi';
 import { PaymentLinkParams } from '@dsh/app/shared/services/create-payment-link/types/payment-link-params';
-import { FormGroupSuperclass } from '@dsh/utils';
+import { WrappedFormGroupSuperclass } from '@dsh/utils';
 
 import { HoldExpiration } from '../../services/create-payment-link/types/hold-expiration';
 import { ORDERED_PAYMENT_METHODS_NAMES } from '../../services/create-payment-link/types/ordered-payment-methods-names';
@@ -29,7 +28,7 @@ import ProvidersEnum = PaymentTerminal.ProvidersEnum;
     providers: [provideValueAccessor(CreatePaymentLinkFormComponent)],
 })
 export class CreatePaymentLinkFormComponent
-    extends FormGroupSuperclass<Controls, PaymentLinkParams>
+    extends WrappedFormGroupSuperclass<PaymentLinkParams, Controls>
     implements OnChanges
 {
     @Input() paymentMethods: PaymentMethod[];
@@ -66,7 +65,7 @@ export class CreatePaymentLinkFormComponent
         this.snackBar.open(this.transloco.translate(isCopied ? 'copied' : 'copyFailed'), 'OK', { duration: 1000 });
     }
 
-    protected innerToOuter({ holdExpiration, paymentMethods, ...value }: ControlsValue<Controls>): PaymentLinkParams {
+    protected innerToOuter({ holdExpiration, paymentMethods, ...value }: Controls): PaymentLinkParams {
         return {
             ...(value.paymentFlowHold ? { holdExpiration } : {}),
             ...value,
