+++ src/utils/form/index.ts
@@ -1,4 +1,5 @@
 export * from './set-form-array-value';
 export * from './get-form-value-changes';
 export * from './get-form-validation-changes';
-export * from './form-group-superclass.directive';
+export * from './wrapped-abstract-control-superclass';
+export * from './wrapped-form-group-superclass';
